Bagslot Bug Fix 2
A bag can no longer be unequipped while it has items inside if the newly equipped bag has fewer slots.
This is also true when the bag is being removed without a new bag taking it's place.
Items are allowed to be inside the bag when upgrading your bag to a bigger one.