Combat Update
+ Added the ability to punch.
When punching, the player will face to where you are aiming with the camera, and will stay positioned like that for a short while.
+ Added a punching animation.
+ Added a walking punching animation.
+ Added enemies.
Enemies can be hit with with punches, and will die when they reach 0 HP. Each punch deals 1 damage.
+ Added damage numbers when hitting enemies.
Whenever you hit an enemy, a damage number will briefly show up, showing you how much damage you have done. The floating number will be tracked to the enemy, and move with it if it moves.